pulling the logic out of the main settings script, and the router, to contain it in 'mainMenu.js' with its own config file, just so that it's less intrusive to the rest of the codebase
