+++ src/routes/index.js
@@ -22,10 +22,6 @@ import Machine from './Settings/Machine.vue'
 import Page404 from './Page404.vue'
 
 
-// menu starts with a safe clone out of the settings state...
-let cachedMenuJsonString = JSON.stringify(store.getters['settings/mainMenuConfig']);
-export const Menu = Vue.observable(JSON.parse(cachedMenuJsonString))
-
 export const Routes = []
 
 Vue.use(VueRouter)
@@ -35,10 +31,6 @@ const router = new VueRouter({
 	base: process.env.BASE_URL,
 	routes: Routes
 });
-
-// when a system component registers, keep the deets so we can re-inject after the settings update
-const keyedMenuItems = [];
-
 
 // Register a new route and menu item
 // component: Component to register
@@ -92,54 +84,12 @@ function registerRouteInternal(component, route) {
 		routeObj.translated = false;
 	}
 
-	injectIntoMenu({ routeObj, routeConfig, namedMenuGroup });
-
-	if (route.key) {
-		// system/internal components have keys, so we can reload them when config is updated
-		keyedMenuItems.push({ routeObj, routeConfig });
-	}
+	// push it into the menu itself...
+	store.commit('mainMenu/setMenuItem', { routeObj, routeConfig, namedMenuGroup });
 
 	Routes.push(routeObj);
 	router.addRoute(routeObj);
 }
-
-/**
- * Main menu starts out by default as categories, this will seek through the menu
- * and place the items under the named groups.
- *
- * The configuration however can be changed, where string placeholders are replaced
- * by actual components when/if they turn up. So in the configuration, place the
- * path for what will be a plugin's path, and when it's injected into the menu,
- * this will fine that string and set th menu item.
- */
-function injectIntoMenu({ routeObj, routeConfig, namedMenuGroup }) {
-	let placed = false;
-	// place into the menu
-	Menu.forEach((item, k) => {
-		if (typeof item === 'string' && (item === routeConfig.key || item === routeConfig.path)) {
-			// replace a string placeholder of top-level button/item (no submenu)
-			Menu[k] = routeObj;
-			placed = true;
-
-		} else if (item.pages) {
-			// place into a submenu
-			let idx = -1;
-			if (routeConfig.key) idx = (item.pages || []).indexOf(routeConfig.key);
-			if (idx < 0) idx = (item.pages || []).indexOf(routeConfig.path);
-			if (idx > -1) {
-				item.pages[idx] = routeObj;
-				placed = true;
-			}
-		}
-	});
-
-	if (!placed && namedMenuGroup) {
-		// backwards compatibility for existing plugins to inject
-		let group = Menu.find(item => item.name === namedMenuGroup);
-		if (group) group.pages.push(routeObj);
-	}
-}
-
 
 export const GeneralSettingTabs = Vue.observable([])
 export const MachineSettingTabs = Vue.observable([])
@@ -171,37 +121,6 @@ export function registerSettingTab(general, name, component, caption, translated
 		MachineSettingTabs.push(tab);
 	}
 }
-
-// Menu config updates, so we need to reset the menu and load
-// it back in as if it was configured that way originally...
-store.subscribe((mutation) => {
-	if (mutation.type === 'settings/load' && mutation.payload.mainMenuConfig && mutation.payload.mainMenuConfig.length) {
-		let tmpString = JSON.stringify(mutation.payload.mainMenuConfig);
-
-		if (tmpString === cachedMenuJsonString) {
-			// nothing changed in the menu, abort updates...
-			return;
-		}
-
-		// quick clear-out of the menu
-		for (const item of Menu) {
-			if (item.pages) {
-				while (item.pages.length > 0) item.pages.pop();
-			}
-		}
-		while (Menu.length > 0) {
-			Menu.pop();
-		}
-
-		// set the new configuration...
-		cachedMenuJsonString = tmpString;
-		let newStart = JSON.parse(tmpString);
-		newStart.forEach(item => Menu.push(item));
-
-		// re-attach the system components
-		keyedMenuItems.forEach(c => injectIntoMenu(c));
-	}
-});
 
 
 // Control
