+++ src/utils/path.js
@@ -162,6 +162,7 @@ const pathObj = {
 	dwcFactoryDefaults: '0:/sys/dwc-defaults.json',
 	legacyDwcFactoryDefaults: '0:/sys/dwc2-defaults.json',
 	dwcPluginsFile: '0:/sys/dwc-plugins.json',
+	dwcMainMenuFile: '0:/sys/dwc-mainMenu.json',
 
 	boardFile: '0:/sys/board.txt',
 	configFile: 'config.g',
